pyproject.toml
[project]
name = "podflow"
version = "0.1.0"
description = "Podcast content intelligence pipeline — detect, transcribe, and catalog podcast episodes"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "feedparser>=6.0",
    "requests>=2.31",
    "assemblyai>=0.33",
    "google-api-python-client>=2.100",
    "google-auth-oauthlib>=1.1",
    "google-auth-httplib2>=0.2",
    "pyyaml>=6.0",
    "click>=8.1",
    "rich>=13.0",
    "pydantic>=2.5",
    "python-dotenv>=1.0",
]

[project.scripts]
podflow = "podflow.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/podflow"]
